Implement case-insentive search for user's publicName
This commit updates the user querying logic to perform a case insensitive search for the `publicName` field. The `findFirst` method in prisma is now utilize with a the `mode` option set to `insensitive`, ensuring that the search for `publicName` doesn't consider letter-casing.

Changes include:
- Modyficatio in the Prisma query to withing the user.service.ts file method `updatePublicName` to include `mode` option set to `insensitive` for `publicName`.
- This ensures that the searches for `publicName` are effectively case-insensitive.

This update aligns with the goal of avoiding a user to have a duplicated `publicName` even if it is a differnce in letter-casing.